Time: 0 ms (100%), Space: 71.7 MB (51.93%) - LeetHub

+++ 0053-maximum-subarray/0053-maximum-subarray.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int maxSubArray(vector<int>& nums) {
-        long long sum = 0;
-        long long maxi = INT_MIN;
+        int sum = 0;
+        int maxi = INT_MIN;
         for (int i = 0; i < nums.size(); i++){
             sum += nums[i];
             if (sum > maxi ){
